将filetitervisitor放到filefilter文件中
