+++ src/main/java/com/github/hcsp/polymorphism/FileFilter.java
@@ -1,9 +1,9 @@
 package com.github.hcsp.polymorphism;
 
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.nio.file.*;
+import java.nio.file.attribute.BasicFileAttributes;
+import java.util.ArrayList;
 import java.util.List;
 
 public class FileFilter {
@@ -22,12 +22,33 @@ public class FileFilter {
      * 实现一个按照扩展名过滤文件的功能
      *
      * @param rootDirectory 要过滤的文件夹
-     * @param extension 要过滤的文件扩展名，例如 .txt
+     * @param extension     要过滤的文件扩展名，例如 .txt
      * @return 所有该文件夹（及其后代子文件夹中）匹配指定扩展名的文件的名字
      */
     public static List<String> filter(Path rootDirectory, String extension) throws IOException {
         FileFilterVisitor visitor = new FileFilterVisitor(extension);
         Files.walkFileTree(rootDirectory, visitor);
         return visitor.getFilterNames();
+    }
+}
+
+class FileFilterVisitor extends SimpleFileVisitor<Path> {
+    private final String extension;
+    private final List<String> filterNames = new ArrayList<>();
+
+    public FileFilterVisitor(String extension) {
+        this.extension = extension;
+    }
+
+    public List<String> getFilterNames() {
+        return filterNames;
+    }
+
+    @Override
+    public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+        if (file.getFileName().toString().endsWith(extension)) {
+            filterNames.add(file.getFileName().toString());
+        }
+        return FileVisitResult.CONTINUE;
     }
 }
